Module Bimage_unix.Magick

Magick contains image I/O operations using ImageMagick/GraphicsMagick on the command-line

val read : ?create: ( string -> ( 'a, 'b ) Bimage.Type.t -> 'c Bimage.Color.t -> int -> int -> ( 'a, 'b, 'c ) Bimage.Image.t ) -> ( 'a, 'b ) Bimage.Type.t -> [< `Gray | `Rgb | `Rgba ] as 'c Bimage.Color.t -> ?format:string -> string -> ( ( 'a, 'b, 'c ) Bimage.Image.t, Bimage.Error.t ) Stdlib.result

read filename kind color loads an image from filename on disk using the given kind and color

val write : ?quality:int -> ?format:string -> string -> ( 'a, 'b, [< `Gray | `Rgb | `Rgba ] ) Bimage.Image.t -> unit

write filename image saves an image to filename

val read_all : ?create: ( string -> ( 'a, 'b ) Bimage.Type.t -> 'c Bimage.Color.t -> int -> int -> ( 'a, 'b, 'c ) Bimage.Image.t ) -> ( 'a, 'b ) Bimage.Type.t -> [< `Gray | `Rgb | `Rgba ] as 'c Bimage.Color.t -> ?format:string -> string array -> ( Bimage.Input.t, Bimage.Error.t ) Stdlib.result

Read multiple images directly into an Input array

val convert_command : string Stdlib.ref

convert_command contains the command used to call out to ImageMagick/GraphicsMagick. For example, if you'd like to use GraphicsMagick then set this to "gm convert"

val identify_command : string Stdlib.ref

identify_command contains the command used to get information about image dimensions. It defaults to itentify * but if you'd like to use GraphicsMagick then set this to "gm identify"